+++ src/main/java/de/maxhenkel/resourcepackchecker/mixin/ClientPacketListenerMixin.java
@@ -1,15 +1,25 @@
 package de.maxhenkel.resourcepackchecker.mixin;
 
+import de.maxhenkel.resourcepackchecker.IFilePackResource;
 import de.maxhenkel.resourcepackchecker.ResourcePackChecker;
+import de.maxhenkel.resourcepackchecker.ShaUtils;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.protocol.game.ClientboundResourcePackPacket;
+import net.minecraft.network.protocol.game.ServerboundResourcePackPacket;
+import net.minecraft.server.packs.PackResources;
+import net.minecraft.server.packs.repository.Pack;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import java.io.File;
+import java.util.Collection;
+
 @Mixin(ClientPacketListener.class)
-public class ClientPacketListenerMixin {
+public abstract class ClientPacketListenerMixin {
 
     @Inject(method = "handleResourcePack", at = @At("HEAD"), cancellable = true)
     private void setServerPack(ClientboundResourcePackPacket clientboundResourcePackPacket, CallbackInfo ci) {
@@ -18,5 +28,28 @@ public class ClientPacketListenerMixin {
             ci.cancel();
         }
     }
+
+    @Inject(method = "handleResourcePack", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/protocol/game/ClientboundResourcePackPacket;isRequired()Z", shift = At.Shift.AFTER), cancellable = true)
+    private void handleResourcePack(ClientboundResourcePackPacket packet, CallbackInfo ci) {
+        if (packet.getHash().isBlank()) {
+            return;
+        }
+        Collection<Pack> selectedPacks = Minecraft.getInstance().getResourcePackRepository().getSelectedPacks();
+        for (Pack pack : selectedPacks) {
+            PackResources resources = pack.open();
+            if (resources instanceof IFilePackResource) {
+                File f = ((IFilePackResource) resources).getFile();
+                String packSha1 = ShaUtils.getSha1(f);
+                if (packSha1.equals(packet.getHash())) {
+                    send(ServerboundResourcePackPacket.Action.ACCEPTED);
+                    ci.cancel();
+                    break;
+                }
+            }
+        }
+    }
+
+    @Shadow
+    protected abstract void send(ServerboundResourcePackPacket.Action action);
 
 }