+++ src/main/java/de/maxhenkel/resourcepackchecker/mixin/DownloadedPackSourceMixin.java
@@ -1,12 +1,12 @@
 package de.maxhenkel.resourcepackchecker.mixin;
 
 import de.maxhenkel.resourcepackchecker.IFilePackResource;
+import de.maxhenkel.resourcepackchecker.ShaUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.DownloadedPackSource;
 import net.minecraft.server.packs.PackResources;
 import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.PackSource;
-import org.apache.commons.codec.digest.DigestUtils;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -14,7 +14,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
 
@@ -26,13 +25,13 @@ public class DownloadedPackSourceMixin {
 
     @Inject(at = @At("HEAD"), method = "setServerPack", cancellable = true)
     private void setServerPack(File file, PackSource packSource, CallbackInfoReturnable<CompletableFuture<Void>> cir) {
-        String serverPackSha1 = getSha1(file);
+        String serverPackSha1 = ShaUtils.getSha1(file);
         Collection<Pack> selectedPacks = Minecraft.getInstance().getResourcePackRepository().getSelectedPacks();
         for (Pack pack : selectedPacks) {
             PackResources resources = pack.open();
             if (resources instanceof IFilePackResource) {
                 File f = ((IFilePackResource) resources).getFile();
-                String packSha1 = getSha1(f);
+                String packSha1 = ShaUtils.getSha1(f);
                 if (serverPackSha1.equals(packSha1)) {
                     serverPack = null;
                     CompletableFuture<Void> future = new CompletableFuture<>();
@@ -41,17 +40,6 @@ public class DownloadedPackSourceMixin {
                     break;
                 }
             }
-        }
-    }
-
-    private String getSha1(File file) {
-        try {
-            FileInputStream fis = new FileInputStream(file);
-            String sha1 = DigestUtils.sha1Hex(fis);
-            fis.close();
-            return sha1;
-        } catch (Exception e) {
-            return "";
         }
     }